+++ tests/integration/test_storage_iceberg_schema_evolution/configs/config.d/named_collections.xml
@@ -1,0 +1,15 @@
+<clickhouse>
+  <named_collections>
+    <s3>
+        <url>http://minio1:9001/root/</url>
+        <access_key_id>minio</access_key_id>
+        <secret_access_key>ClickHouse_Minio_P@ssw0rd</secret_access_key>
+    </s3>
+    <azure>
+        <account_name>devstoreaccount1</account_name>
+        <account_key>Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==</account_key>
+    </azure>
+    <local>
+    </local>
+  </named_collections>
+</clickhouse>
