+++ tests/integration/test_storage_iceberg_schema_evolution/configs/users.d/users.xml
@@ -1,0 +1,9 @@
+<clickhouse>
+    <users>
+        <default>
+            <password></password>
+            <profile>default</profile>
+            <named_collection_control>1</named_collection_control>
+        </default>
+    </users>
+</clickhouse>
